Properly set sound notification (fixes #117)
Set the ringtone URI to the default ringtone if sound notification is
enabled. Set it to null otherwise.

This fixes the problem that sound notifications are not sent anymore
once the user has disabled and then re-enabled them. The reason was,
that the ringtone URI would never be set to a valid value once it had
been set to null.