Fix SMP for non-ASCII chars.
secret.length() is the length of the String, but OTR needs to be aware of the
length of the actual byte[]. With UTF-8, the length of the string will be
different than the number of bytes.  This is easily confused because with
ASCII, the length of the String and the byte[] will be the same.

closes https://github.com/otr4j/otr4j/pull/47