remove explict cursor.close() where not necessary
there is an error that is constantly being thrown
but that doesn't cause the app to quite, related
to finalize() being called on a cursor that was closed,
or something of that nature. It is not necessary
to explicitly call close() on all cursors, so in these
case we will not.