initial fixes to reduce SQL injection attack risks
though unlikely, it is possible for an app to use intents
to cause a local, on-device SQL injection attack.

in addition, via a chat message, a remote user could also
cause a sql injection.

we are working on changes to improve parsing and use of
parameterized inputs to all queries