ensure OtrAndroidKeyManagerImpl is created whenever getInstance() is called
Instead of passing in the central keystory File in one place, and a null in
another, pass in the Context in both places, and move the filename into
getInstance() since that's the only place that ever needs it.  Otherwise if
the OtrAndroidKeyManagerImpl instance does not exist yet, and the arg is
null, getInstance() returns _instance which is still null.