fail fast on failure to initialize OTR
OTR is not optional.  Failure to initialize OTR makes incoming messages drop on the floor.

This could happen if the OTR keystore is corrupt, the password is incorrect or there is a bug
in order of initialization.  I suspect the latter in a failure case I saw with one of my testers.