on Sign In, put indeterminite, spinning progress meter in ActionBar
There is not a lot of feedback once you hit the "Sign in" button, so this
tries to put a spinning progress meter in the ActionBar once the "Sign in"
button is pressed.  Because a lot of work is happening on the main thread,
this progress meter does not always show up very quickly.  Ideally, all of
the actions triggered by pressing "Sign in" would happen in an AsyncTask,
and AccountActivity would finish once the AsyncTask finishes.