guarantee a unique XMPP Resource
Since XMPP Servers can direct traffic based on the XMPP Resource,
ChatSecure should ensure that each install has a unique XMPP Resource.
Otherwise, if someone has two instances of ChatSecure running on the same
account, the XMPP Server might always direct traffic to one but not the
other.  Since ChatSecure does not expose the XMPP Resource to the user, it
should be a random string that is then stored

refs #563 https://github.com/guardianproject/ChatSecureAndroid/pull/563
refs #520 https://github.com/guardianproject/ChatSecureAndroid/issues/520