send OTR "FINISHED" message to all OTR sessions on logout
If the user logs out of an account, send the OTR message to end the OTR
session, which notifies the client on the other end that it is no longer
possible to guarantee that their messages will be sent encrypted (e.g. if
the message gets queued as plain text and then resent before OTR is
renegotiated).

refs #520 https://github.com/guardianproject/ChatSecureAndroid/issues/520