tweak displayed xmpp: URL to work on more apps and follow RFC5122
The format of the xmpp: URL in the QR Code did not follow the standard, and
therefore only worked in ChatSecure-Android.  Chris Ballinger had previously
thought that xmpp:to@foo.com URLs didn't work in iOS but he has since gotten
them working.  Therefore it makes sense to use a fully proper RFC5122 URL to
advertise in ChatSecure.

refs #4092 https://dev.guardianproject.info/issues/4092