The locale folder "id" should be called "in" instead
From the java.util.Locale documentation:
"Note that Java uses several deprecated two-letter codes. The Hebrew ("he")
language code is rewritten as "iw", Indonesian ("id") as "in", and Yiddish
("yi") as "ji". This rewriting happens even if you construct your own Locale
object, not just for instances returned by the various lookup methods.

Because of this, if you add your localized resources in for example values-id
they will not be used, since the system will look for values-in instead.

To work around this, place your resources in a values folder using the
deprecated language code instead.

https://developer.android.com/reference/java/util/Locale.html