check status of pref.commit() otherwise just use pref.apply()
apply() does it in the background, but does not provide confirmation.
commit() provides confirmation, but is only useful if the return value is
checked.  apply() happens in the background and commit() in the foreground,
so using apply() should keep things more interactive. The SharedPreferences
docs recommend using apply() if the return value of commit() is ignored.