show error message if sharing data when all accounts are offline
Previously, if a SEND/Share Intent was sent to ChatSecure when it was
offline, ChatSecure would just show a blank white screen.  That screen was
ImUrlActivity doing nothing at all.  When all accounts are offline, this
commit makes the blank white screen go away, and gives an error message in
a Toast.