when user selects "End chat", also send OTR FINISHED message
The "End chat" is a very distinct action that is not often taken.  It also
deletes the existing history, etc. so the expectation is that it sets
things up to be more secure.  Therefore, it should also send the OTR
FINISHED message.  That signals the other side of the conversation that OTR
is no longer available, so no message encryption will happen until a new
OTR session is negotiated.  That will prevent the user from mistakenly
sending a message that goes through unencrypted.