only single background image file needed, its dynamically resized
As far as I understand the code from Google's recommended approach for loading
large background images, it is reading a resized version into memory, so there
isn't really a need for different sizes of the background image.  It would
only save a little CPU time, which doesn't seem worth the added complexity.

refs #4277 https://dev.guardianproject.info/issues/4277

This reverts commit 0a040b26d2d21f45fd6066ef316e53c4f81ebc3a.
             "add new low-res versions of default background"