Update update-ant-build.sh to work correctly on current build tool versions
If you check out the project from scratch and run the
update-ant-build.sh script with current Android build tools, it fails
with the message 'Error: The project either has no target set or the
target is invalid.', as there is no target parameter in the android
update call. The bat file does not make this mistake.