make-release-build: generate APK file name based on `git describe`
This marks the APK with the actual version, not just the version that is in
AndroidManifest.xml.  If this build is on a signed tag that matches the
versionName in AndroidManifest.xml, then the APK will be a proper release
APK.  Otherwise, it'll get the whole `git describe` string to mark it as an
non-release build.