support new IOCipher v0.3 API
With IOCipher v0.3, you have to explicitly create a new VFS file, whereas
before VirtualFileSystem.mount() would do it automatically.  This made it
hard to debug situations where something went wrong with the file at the
path that was stored, like the perms changed, or the file was deleted by
another process.  This change makes it more like the UNIX mount command.

refs #4531 https://dev.guardianproject.info/issues/4531