use a ScrollView for Account Wizard screens to accomodate long translations
Some languages are more verbose, or also have longer spellings.  Long
languages on a small screen mean the buttons get obscured.

This looks like a big change because I also formatted the file with the
Android formatter.  But literally I just changed the most external item
from a LinearLayout to a ScrollView.

closes #4540 https://dev.guardianproject.info/issues/4540