store the location of IOCipher media store so if its missing, prompt user
Right now, the IOCipher container file that is the encrypted media store is
not stored in a stable path.  If the External Storage is present when
ChatSecure starts, then it is used, otherwise the Internal Storage is used.
But if ChatSecure starts, then the user ejects the External Storage,
ChatSecure will just try the External Storage path, then crash when it is
not there.  Or vice-versa: if ChatSecure is started when the External
Storage is not mounted, then the user mounts the External Storage,
ChatSecure will look for the media store on External Storage, and crash.

The first step to fixing this problem is storing a preference of where the
media store file should be, so ChatSecure can know whether to just create a
new media store because there never was one, or whether to prompt the user
to put the External Storage back in.  To do that, we have to detect all the
possible variations that could have happened, so there is some crazy logic
in IocVfs.init().