provide single method for "Shutdown & Lock"
This creates a single static method for calling "Shutdown & Lock", so that
other Activities can call it without recreating this code.  Specifically,
for the error dialog when the IOCipher media file is missing or the SD Card
is not present.

This probably should instead be implemented as a static method on ImApp,
but this would require changing how CacheWord works, and that's a bigger
project than should be tackled right before a release.

This follows up on the previous commit, which removed a dead version of
this same method.