create prompt screen if chat media store or SD card is missing
The media store was sometimes stored on the External Storage, which can be
removed or even replaced with a different SD Card.  This adds a prompt
screen at startup to warn the user about this, and give them two options
for fixing the situation: delete references to the existing media store or
try changing SD cards.

closes #4531 https://dev.guardianproject.info/issues/4531