stop delete prompt after taking audio/pictures, just delete automatically
This removes the Dialog that asks the user whether they want to delete the
original audio or picture file that they just took from ChatSecure in
order to send in a message.  There is no need for this Dialog and it is a
rather big annoyance in a "push to talk" session. If the user wants to
save an image or audio file, they'll first want to see or hear it, which
they can do once it is displayed in the message view.  If they want to
keep it, they can long-press it and choose "Export".