purge old code structure that launches NewChatActivity for each account
Based on the method names, it looks like this code structure is left over
from when things behaved quite a bit differently.  For example, launching
NewChatActivity is the sole action of a method called showAccounts(), and
showActiveAccount() iterates through all active accounts and launches a new
Intent to show NewChatActivity.