make sure all desired accounts sign in on startup
There was a bug on start up when there are multiple accounts set up.
Something is setting one of the accounts as already signed in, i.e.
Imps.ConnectionStatus.ONLINE before signInAll() is run.  It could also be that
those values were not updated when ChatSecure shutdown, or ChatSecure was
killed before it could update those values.

Really, the offline/connecting/online status info should be fetched live from
XMPP/SMACK rather than using some stored value in the database.  The only such
value that should be stored is ACTIVE_ACCOUNT_KEEP_SIGNED_IN to see whether
the user wants that account to start up automatically. Right now there are
four very similar values stored, but not really consistently used:

* Imps.Account.LAST_LOGIN_STATE
* Imps.Provider.ACCOUNT_CONNECTION_STATUS
* Imps.Provider.ACTIVE_ACCOUNT_KEEP_SIGNED_IN
* Imps.AccountStatus.CONNECTION_STATUS

The particular bit of code changed in this commit only seems to be used when
ChatSecure is setup with a password.  With ChatSecure is used without a
password, the startup stuff is handled in RemoteImService.autoLogin().