make-release-build: use version from `git describe` in original name format
This makes it so that the 'v' is included before the version, like the
ChatSecure releases, i.e. ChatSecure-v14.2.0.apk rather than
ChatSecure-14.2.0.apk.

For builds of versions that are not tagged, the version name will be the
format of `git describe`, which includes the previous tag and commit ID.