make-release-build: disable faketime since it is problematic
faketime is used to normalize all of the timestamps to get a reproducible
build.  Since this is still far off from getting a perfectly reproducible
build, i.e. the same hashsum, just remove this for now.  faketime is not
needed to get a reproducible build in terms of the APK signature, only the
hash.